drf/requirements.txt
absl-py==0.9.0
appdirs==1.4.4
asgiref==3.2.10
astor==0.8.1
astroid==2.4.2
astunparse==1.6.3
attrs==19.3.0
audioread==2.1.9
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.1.5
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
branca==0.4.1
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
Django==3.1
django-cors-headers==3.5.0
django-filter==2.4.0
djangorestframework==3.12.1
djangorestframework-jwt==1.11.0
docopt==0.6.2
docutils==0.15.2
entrypoints==0.3
et-xmlfile==1.0.1
eunjeon==0.4.0
Flask==1.1.2
folium==0.11.0
gast==0.3.3
gensim==3.8.3
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.30.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel==5.3.0
ipython @ file:///C:/ci/ipython_1593433781269/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jdcal==1.4.1
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
jmespath==0.10.0
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
JPype1 @ file:///C:/Users/Lagom/lagom/JPype1-0.7.5-cp37-cp37m-win_amd64.whl
Js2Py==0.70
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kerascv==0.0.40
kiwisolver==1.2.0
konlpy==0.5.2
krwordrank==1.0.3
lazy-object-proxy==1.4.3
librosa==0.8.0
lightgbm==2.3.1
llvmlite==0.34.0
lxml==4.5.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mglearn==0.1.9
missingno==0.4.2
mistune==0.8.4
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.5.0
music21==6.1.0
nbconvert==5.6.1
nbformat==5.0.7
nltk==3.5
notebook==6.0.3
numba==0.51.2
numpy @ file:///C:/ci/numpy_and_numpy_base_1603468620949/work
oauthlib==3.1.0
opencv-python==4.4.0.40
openpyxl==3.0.5
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pickleshare==0.7.5
Pillow==7.2.0
pipwin==0.5.0
plotly==4.9.0
pooch==1.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.12.4
pyarrow==1.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/Lagom/pipwin/PyAudio-0.2.11-cp37-cp37m-win_amd64.whl
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
pygame==1.9.6
Pygments==2.6.1
pyjsparser==2.7.1
PyJWT==1.7.1
pylint==2.5.3
pyparsing==2.4.7
PyPrind==2.11.2
pyrsistent==0.16.0
pySmartDL==1.3.4
PySocks==1.7.1
python-dateutil==2.8.1
python-rake==1.4.5
pytz==2020.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
rake-nltk==1.0.4
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
retrying==1.3.3
rsa==4.6
s3transfer==0.3.3
scikit-learn @ file:///C:/ci/scikit-learn_1592847564598/work
scikit-surprise==1.1.1
scipy @ file:///C:/ci/scipy_1592916958183/work
seaborn==0.10.1
Send2Trash==1.5.0
six==1.15.0
smart-open==2.1.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
tensorflow-hub==0.8.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.1
tornado==6.0.4
tqdm==4.47.0
traitlets==4.3.3
tweepy==3.9.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
tzlocal==2.1
urllib3==1.25.9
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webcolors==1.11.1
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wordcloud==1.7.0
wrapt==1.12.1
xgboost==0.90
xlrd==1.2.0
zipp==3.1.0
